fix: CombinedStockAnalysisSchema의 close 필드를 int에서 float로 변경

+++ dolpha/schemas.py
@@ -102,7 +102,7 @@ class CombinedStockAnalysisSchema(
     FinancialDataSchema, StockAnalysisSchema, CompanySchema
 ):
     # OHLCV 추가 필드 (상승률 정보용)
-    close: Optional[int] = None
+    close: Optional[float] = None
     change: Optional[float] = None
     # 52주 최저가 대비 상승률
     min_52w_gain_percent: Optional[float] = None
